FEM: geom tools, add new module and move some geom tools from mesh tools in
